feat: LC repeat_count가 10에서 멈추는 버그 조치 (draft 생성 로직 수정) / ai 분석 오류 해결 / 설계 변경으로 incident published status 삭제 조치

+++ src/main/java/com/soyunju/logcollector/repository/lc/ErrorLogHostRepository.java
@@ -96,6 +96,6 @@ public interface ErrorLogHostRepository extends JpaRepository<ErrorLogHost, Long
     List<com.soyunju.logcollector.repository.lc.agg.HostAgg> findTopHashesByHostCount(@Param("limit") int limit);
 
     @Query(value = "SELECT COUNT(*) FROM error_log_hosts WHERE log_hash = :logHash", nativeQuery = true)
-    long countHostsByLogHash(@Param("logHash") String logHash);
+    int countHostsByLogHash(@Param("logHash") String logHash);
 
 }
